Set internal fontname to font specification
LaTeX expects that fontname is the tfm name, and thus use it to recall
the font, so we ought to store something that can be used the same way
here. `specification.specification' is the string used to define the
font.

Test:

\input luaotfload.sty
\font\xxx=file:lmroman10-regular.otf:+liga;+tlig\xxx
``\fontname\font'' should be: {\tt file:lmroman10-regular.otf:+liga;+tlig}
\bye

closes #3