Improving style guessing based on real-life fonts
Now all the informations are printed in the guessed_style, for parsing
facilities. Light and Heavy are not treated as weight anymore (some
fonts are heavy regular or heavy bold in the same family). Treating the
condensed style separately. If it could be tested on a large database it
would help me improve it. Of course perverse styles like the ones of
semaphor are not guessed, but they are not usual styles so it's not a
problem.