Port to "alt_getopt"
Now the options are handled the standard getopt_long way, with a hack to
support multiple -v's, and --verbose now takes a required level
argument.

Also, we now use require() which already have built-in error checking
(if it doesn't work on some system, then this is a bug to report).