Fix optical size matching
With the previous commit, if we don't found a font with the requested
optical size, we return nothing, though we might have a matching font
name.

Now we store all matching names with no matching optical size and return
the first one, then we return the closest match (i.e. if the requested
size is 15pt and we have 10pt and 5pt match, we return the 10pt one).