Fix loading subfonts
Previously, we relied in the fact that origname contains the subfont in
the form of |foo.otf(subfoo)| and that luatex would then load the right
font, but this is not what the rest of the code expects and cause some
bugs. We now return the filename and subfont properly (i.e.separately).