Make sure the font exists on the disk
Sometimes the database is outdated in the font no longer exits in the
saved path, now we make sure the file exists before returning path to
luatex. This way the user gets the good old "font not loadable" error,
instead of mysterious backend error and program exit. Later we may try to
update the database at this point.