reduce verbosity of the font loader
this is my version of:

    https://github.com/phi-gamma/luaotfload/commit/3a59b8b6b226ed22ccd16aec2dc2e272f015cc74

I undid the original because our policy is not to change imported files
except for applying hot fixes in advance that Hans will add eventually.

However, we’ll see if we really need otfl-fonts.lua anyways as its
functionality might fit better in to luaotfload.lua.