remove old fond definition callback
everything it did is already been done by
``constructors.assignmathparameters()`` in ``font-con.lua``, so it is
save to remove the old code without replacement.

There is one subtle difference, however: Context scales the value of
``FractionDelimiterDisplayStyleSize`` by a factor of 2.40, whereas the
old luaotfload used 2.39. I’d say whoever takes offense because of this
difference please debate Hans for the correct value. kthxbye