make font name rewriting optional
... but still activated by default

the trouble with rewriting the font name is that it diverges from the
behavior as stated in the Luatex reference. according to chapter 7, the
field ``name`` of the font structure should contain the “metric (file)
name”. by including the complete specification we cater to hacks that
feed ``\fontname`` (which reads the ``name`` field) back into ``\font``
as Xetex does. this appears to work without complications, but formats
other than Latex might not need it. so the best strategy would be to
make it an option and have Fontspec activate this non-standard behavior.