[letterspace] convert to node.direct
This is an experiment: The character kerning callback has been converted
in its entirety to utilize the not-so-new direct node API. Since the
translation was mechanical only to a certain extent, this may have
introduced errors. On the other hand, the revised code resembles its
distant ancestor in Context’s ``typo-krn.lua`` much more closely again,
which may come in handy in the future.