[mkstatus,build,conf] default to packaged fontloader via status file
This adds an entry ``loader`` on the ``notes`` hash in the status table
by means of which the fontloader can be specified. The status file is
read when building the configuration table, and if it is present the
entry in the ``loader`` table will be used as the default fontloader.
Otherwise, Luaotfload falls back on the reference loader.