[init,main,package] split initialization in early and late stage
This is a big move. Due to the restructuring of the initialization we
can now separate the early tasks -- mostly dirty stuff like writing
globals -- from the later stage where we actually inject the actual font
loader. As a consequence, the basic configuration is now present before
the font loader is injected. It’s not complete, though, in that the
reconfiguration step must still be executed after the font loader is
there, since some config variables require values to be set up by the
loader. A decision is pending regarding how we address the situation;
possible approaches are, in order of increasing complexity:

    * reconfigure twice, skip some of the steps the first time through
    * split reconfiguration into two stages
    * make configuration independent of fontloader values

Design-wise, the last point would be the most desirable, naturally.
Though the fontloader has been known to require extensive static
preparation which might very well make that strategy impossible or
unjustifiably demanding. We’ll see how it plays out …