Patch to not assume luatexbase needs to be loaded
The ltluatex code being trialed at

https://github.com/josephwright/ltluatex

includes a luatexbase emulation package that allows
luaotfload to run without change but the core code that
is intended to be included in future latex formats requires
two small changes and also would require that luaotfload
detect that luatex support is already provided and so not
load luatexbase.

\RequireLuaModule is not defined by default in TeX and
luatexbase.reset_callback is not defined by default in lua
in the core ltluatex code.