[db] fix search path construction on Win hosts
This seems to help with the examples from #165 that rely on setting
``$OPENTYPEFONTS`` via the environment.

Referencing a non-existing value caused a branch to always pick Unix
path separators. The affair was pretty insidious since the paths we
received from kpse were sane so the behavior was only triggered with
manual overrides.