[build] kill off install target
A while back we had one confused user who apparently had a borked
install due to this. Since it hasn’t been tested, we replace the
original instructions with a help message.

The target may be reintroduced at some point (build TDS, then extract to
local tree), but for now it’s not worth the hassle.