[fontloader] revert features handler temporarily
Address issue #318

This reverts the changes made to font-otn.lua with commit adc43c3c24..

Certain features related to directionality are not handled correctly
with the most recent fontloader. As a temporary measure, revert the
handling code in font-otn.lua and restore the values it references.
While this appears to suppress the issue, we’ll switch back to the
upstream code once we have an actual solution.