[features,parsers] implement font fallbacks
Building on the combination mechanism, this allows defining fallback
fonts of which all glyphs are pulled that aren’t currently part of the
base font. Example:

    \input luaotfload.sty
    \font \lm  = file:lmroman10-regular.otf:mode=base
    \font \cmu = file:cmunrm.otf:mode=base
    \font \lmu = "combo: 1->\fontid\lm; 2->\fontid\cmu,fallback"
    \lmu Eh bien, mon prince. Gênes et Lueques ne sont plus que des
         apanages, des поместья, de la famille Buonaparte.
    \bye

This allows setting Latin Modern text that contains Cyrillic letters.

Note that -- as with the other combinations -- only glyphs are
considered, no other properties of the fallback font. So besides the
occasional letter in a different script this functionality is probably
useless.