[import] adapt to grand upstream restructuring
Two not quite unrelated changes have to be accounted for:

    * Finally, the source for Context MkII and MkIV are being kept in
      separate trees, at least as far as the format is concerned. All
      out scripts that pull code from upstream had to be adapted.

    * New fontloader. This has been cooking for half a year maybe: Some
      of our troubles with 2.6 were caused by separating out the revised
      but highly experimental code from the self-contained fontloader,
      despite Hans’ great efforts to make it seamless. Now that the new
      fontloader is deemed mature enough, we have to accomodate the
      switchover.

The fallout exceeds the bounds of Luaotfload proper: The Lualibs package
had to be extended as well, adding a rather tiny library from the
``util-*`` namespace. With the latest setup, a fontloader can indeed be
assembled automatically again. No testing has taken place yet, so it
might not even run.