[aux] fix crash in patching code that sets \fontdimen8
Obviously, since Fontforge has been ditched, we need to adapt to the
slightly different data structures created by the Lua reader. For the
time being, we revise the code so it will not crash instantly due to the
lack of a missing ``pfminfo`` table.

Hans has been notified of our use of the ``capheight`` data and may
add that value grudgingly again.