[db,loaders] clarify support for PFB
The current PFB loader, although it is indeed completely independent of
the FF libraries, is not yet feature complete. Only the loading of
vectors is supported which suffices for font rendering given the AFM
information.

According to Hans, we have decent chance of it growing into a
full-fledged reader for 1.0.