[tool,resolvers,conf] switch --find lookups to the actual resolvers
The arguments to “--find” on the command line avoided calling the real
index API functions and used crude approximations instead. In order to
make “--find” obey the new “anon-sequence” configuration item, it needs
to access the normal resolvers instead. This requires certain
adaptations to allow for a fallback on the “file:” lookup.