[main,tool] reintroduce lost version check
The version check was removed by accident in commit 5ce60cc98a30. This
isn’t normally a problem except when people start to run the Pretest
version on ancient Luatex binaries … In any case, the check must be
present. While we’re at it, make the error messages consistent between
the tool and a live TeX run. This can’t cover the fact that there is no
(direct) means of determining the Luatex version when not running in TeX
mode, so the checks still aren’t equivalent.