[aux] fix units lookup prevent crash with AFM fonts
With the new loader, the ``units_per_em`` field resides under the
toplevel TFM structure with the key ``units``. For AFM fonts, it is not
fount under the metadata table too, which we are currently querying.
Thus we prefer the main value, falling back on metadata only in case it
is missing.

At this occasion, tidy up our unit lookup helper and use that wherever
we need access to the values ourselves.