[features,fontloader] fix adding features
This makes our own features work with otf.addfeature().

However, for TFM it’s still not functional out of the box since the
enhancer is installed at a time when it has not been defined yet. An
answer from Hans is pending. In the meantime, we put a crude hack into
our font-tfm.lua to allow injecting the enhancer retroactively on
Luaotfload init.