[db] tentatively reintroduce fallback for italic
Fix assignment of LM series fonts. Currently these are broken because of
borked typosub identifiers like “8oblique” that prevent exact name
matching and at the same time exclude matching the (usable) subfamily.
Introduce a heuristic based on the italic angle value that assigns
italic as a fallback in these cases.

Test: https://bitbucket.org/phg/lua-la-tex-tests/src/857c83ca98cb35153979a0613d3a742bfd93f834/lua/tla-names-3-lm.lua