Fix using HarfBuzz-loaded text fonts in math mode
The simplest fix is to just suck it and load all supported characters in
the font structure.

A more involved fix would to hook the “mlist_to_hlist” callback, look
for glyph noads using HarfBuzz-enabled fonts and cal
font:get_nominal_glyph on them, but this is likely to be slower than the
simpler solution.

Fixes https://github.com/khaledhosny/harf/issues/33