No more harf-base
First this is trying to avoid messing up the luaharfbuzz table. It is a
built-in library, we should not modify it.

Also texlig/texrep where only used once anyway, so they can be made
local. We really don't want people to mess with them.

Also moves CH_GID_PREFIX to a font specific gid_offset which happens to
be constant. The font table is available anyway wherever this is used,
so might actually be faster and it allows to have different values in
different fonts in the future. (E.g. someone might want to use valid Unicode
codepoints in the PUA for fonts where these are available)
Also cahnges the nameing from PREFIX to offset. No deep reason for that,
it just feels more natural. The value changed from 0x110100 to 0x120000
to make it easier to see the GID if the value is given in hex. Also it
allows to add additional glyphs at additional planes.