Use `kpse.show_path` for getting search paths
`kpse.show_path` is more accurate if we want a search path for a
particular file type. Kpathsea then handles by itself what variables to
use. However, further processing of fontdirs needs the path fully
expanded, which requires combination of `show_path` and `expand_path`.

Additionally this works better for the case when variables associated to
file types (e.g. $TTFONTS) are not set anywhere (neither in the
environment, nor in configuration file), because compile time default
paths are used.

Checking for emptiness of `fontdirs` doesn't handle the case when all
the expansions are empty. The final value of `fontdirs` would be ':::'
(or ';;;') and that slips the emptiness check. This is now handled by
checking the result of `filesplitpath`, so even the unlikely case when
all the compile time defaults are empty is covered.