Further bugfixing
The situation this and c653f9ca39 are dealing this are not nominal,
appear in a discretionary and then get \unhbox'ed. These do not get
skipped by the itemization code because that is treating discretionaries
as a atomic blackbox. Therefore we can not assume that all glyph nodes
which are ot skipped have the right font and aren't processed yet.