Fix string formatting of version number
According to Lua 5.3 manual %q is for formatting strings and escaping
them so that they can be read again by Lua. This particular number is
however floating point and %q chooses to format it like with %a, that
means hexadecimal. This greatly reduces the the readability. Because
ConTeXt's OpenType loader versions are always in the format `x.yyy` we
can use %.3f directly.

This is the difference in output:
```
luaotfload | init : Context OpenType loader version 0x1.8e76c8b439581p+1
luaotfload | init : Context OpenType loader version 3.113
```

There may be other places where this issue may be relevant, but this
line is printed to every log file.