Replace non-breaking space by ordinary space
The space between braces on the changed line was actually a non-breaking
space ("c2 a0" in UTF-8). In normal Lua such byte sequence is illegal in
that place.

But in LuaTeX it is actually interpreted as an identifier -- in this
case a global variable evaluating to "nil" and the resulting table being
"{nil}".